% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_game.R
\name{eval_game}
\alias{eval_game}
\title{Eval the overall score of a game}
\usage{
eval_game(g)
}
\arguments{
\item{g}{a \link{game} object.}
}
\value{
(int) the overall score of g
}
\description{
Eval the overall score of a game
}
\examples{
t1 <- turn(1)
g1 <- game(t1)
eval_game(g1) # 1
}
