% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{eval_score_turn}
\alias{eval_score_turn}
\title{Evaluate a complete score for a turn}
\usage{
eval_score_turn(game, n_turn)
}
\arguments{
\item{game}{a \link{game}}

\item{n_turn}{(int) number of turn to evaluate the score (between
1 and 10)}
}
\value{
(int) the score for the \code{n_turn}th turn of the game
}
\description{
Evaluate a complete score for a turn
}
\examples{
game <- game(turn(5, 5), turn(2, 5))
eval_score_turn(game, 1) # 12
}
