% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{game}
\alias{game}
\title{Create game object}
\usage{
game(...)
}
\arguments{
\item{...}{turns to include in the game}
}
\value{
an object of class \link{game}
}
\description{
A game is a sequence of \link{turn}s.
}
\note{
The order of the turns matter!
}
\examples{
t1 <- turn(1, 2)
game(t1)
game(t1, t1)
}
