% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{print.game}
\alias{print.game}
\title{Print method for turn object}
\usage{
\method{print}{game}(x, ...)
}
\arguments{
\item{x}{an object of class \link{game}}

\item{...}{further arguments passed to or from other methods
(currently unused)}
}
\value{
x invisibly
}
\description{
Print method for turn object
}
