% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turn.R
\name{print.turn}
\alias{print.turn}
\title{Print method for turn object}
\usage{
\method{print}{turn}(x, ...)
}
\arguments{
\item{x}{an object of class \link{turn}}

\item{...}{further arguments passed to or from other methods
(currently unused)}
}
\value{
x invisibly
}
\description{
Print method for turn object
}
