% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdojo.bowling-package.R
\docType{package}
\name{rdojo.bowling-package}
\alias{rdojo.bowling}
\alias{rdojo.bowling-package}
\title{rdojo.bowling: Bowling kata for R-TDD Dojo}
\description{
{rdojo.bowling} implements and documents the
    bowling kata for the TDD dojo in R step-by-step.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CorradoLanera/rdojo.bowling}
  \item Report bugs at \url{https://github.com/CorradoLanera/rdojo.bowling/issues}
}

}
\author{
\strong{Maintainer}: Corrado Lanera \email{corrado.lanera@gmail.com} (\href{https://orcid.org/0000-0002-0520-7428}{ORCID})

}
\keyword{internal}
