% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turn.R
\name{turn}
\alias{turn}
\title{Turn}
\usage{
turn(x, y = 0)
}
\arguments{
\item{x}{(int) score for the first roll in the turn}

\item{y}{(int, default 0) score for the second roll in the turn}
}
\value{
an object of class \link{turn} reporting:
\itemize{
\item the sum of the rolls
\item the original rolls
\item the type of turn (standard, spare, or strike)
}
}
\description{
Turn
}
\examples{
turn(1, 1) # 2, standard
turn(1)    # 1, standard
turn(5, 5) # 10, spare
turn(10)   # 10, strike
}
